split footprinter conversion helpers
