+++ cli/convert/convert-options.ts
@@ -1,0 +1,6 @@
+export interface ConvertOptions {
+  footprinter?: boolean
+  json?: boolean
+  name?: string
+  output?: string
+}
